Fixed issues with objects not rendering at certain angles
Implemented a solution from stack overflow to get the true difference between any two angles.
The layering issue persists.
Placement is slightly off, due to objects being placed on whole value points in the grid rather than the middle of grid squares.
Slight rendering issue when you get within a certain range of the circles. I believe this is due to projection, may need to fix this, depending on how it interacts with picking up the objects when you run into them.